Avoid stale log files.
When the logging configuration changes, the logging-related data
structures on the server are not properly updated. This leads to a bug
where logs are written to old files, and it is impossible to fix without
restarting the server. The log file grows indefinitely and eventually
makes the server run out of disk space (see #8696). To avoid
catastrophic consequences, require that the server is restarted to apply
logging configuration changes, until the proper fix is developed.