Remove outdated cmake documentation
When it had been added, initially, it was autogenerated - #14711
Later, after documentation rework, it had been removed, in #37441.

And this makes documentation out dated, and out dated documentation may
provide more harm then help, so let's remove it.

Also it has links to the code in github that does not contain commit
SHA1, so those links may point in the wrong place.

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>