Update docs for max_replica_delay_for_distributed_queries (#40174)
A replica is unusable if its lag is greater than or equal to the
value of max_replica_delay_for_distributed_queries. However, the
current documentation says that a replica is only unusable if it lags
more than the set value.

Also, when the value of max_replica_delay_for_distributed_queries
is read, if it is 0, then a replica is considered up to date
right away without checking its delay. This behavior should be
documented.