tests: avoid model overlap for obfuscator
In case of stress tests it is possible, and there is LOGICAL_ERROR in
case of error, which will create core dump.

Actually on CI [1] there error was likely this:

    stress_test_run_17.txt:/usr/share/clickhouse-test/queries/1_stateful/00175_obfuscator_schema_inference.sh: line 18: /tmp/clickhouse-test/1_stateful/model.bin: No such file or directory

So the file had been removed by another concurrent test.

  [1]: https://s3.amazonaws.com/clickhouse-test-reports/42190/7fb74f05c652f5821d648ddf9b41cc251b89b717/stress_test__ubsan_.html

Note, that actually it is enough just to change the name in this two
tests, however let's make them even more error-resistant.

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>