Do not include libcxx library for C
If you will have the same header in both libraries you may have some
tricky failures.

And there is at least one header that exists in both: stdatomic.h

Right now there is a workaround for this header, that allows to use C++
version of this header for C - the workaround is called "set
_LIBCPP_COMPILER_CLANG_BASED not only for CXX" [1] and use include_next
[2].

  [1]: https://github.com/ClickHouse/libcxx/blob/9a457ab3c64a533a06922b386b284215c17ce627/include/__config#L39
  [2]: https://github.com/ClickHouse/libcxx/blob/9a457ab3c64a533a06922b386b284215c17ce627/include/stdatomic.h#L223-L231

However #42730 changes this, and now it fails to compile because of this.

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>