[TEST] Tune allocator
Processing of the default max_block_size can be faster then running
mmap()/munmap() plus memory dependencies.

Here is an example:

    SELECT count() FROM zeros(10_000_000) WHERE NOT ignore(randomString(1000)) SETTINGS function_implementation='avx2'

- Before this patch it takes: ~6sec
- After: 1.3sec

And even though 128MiB should be enough, since for this query size of
allocation for string will be 65409*(1000+1)=65474409 bytes, due to
rounding to power of two it will not, so let's try simply use 256MiB
(another option is to use strict comparison for MMAP_THRESHOLD) and see
the perf tests.

But also note, that this has other allocator side effects (performance,
fragmentation), so unlikely this is for upstream.

I've found this while I was playing with PODArray [1].

  [1]: https://s3.amazonaws.com/clickhouse-test-reports/45654/2101b66570cbb9eb9a492afa8ab82d562c34336b/performance_comparison_[1/4]/report.html

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>