ActionsDAG: do not change result of and() during optimization
Previously `and` results were inconsistent.
Some optimizations can try to simplify and completely remove a
single-argument `and`, replacing it with a cast to UInt8.
This works well for all numeric types except floats.
A float passed to `and` is normally interpreted as `bool` and
values between 0 and 1 are treated as 1.
The same value cast to an integer type will be truncated to 0.