Inline SipHash::get128(char *)
The SipHash-128 values of the strings are calculated and stored in
a set when the UniqExact aggregate function is applied to a Column-
String, as the following snippet shows:

StringRef value = column.getDataAt(row_num);
UInt128 key;
SipHash hash;
hash.update(value.data, value.size);
hash.get128(key);

The implementation of SipHash requires an internal 8-byte buffer,
current_word (or current_bytes), for holding the input data in the
cross function calls of SipHash::update and SipHash::get128. But,
in the above situation, where the update and get128 only take place
once, the copy of input data to the buffer is unnecessary.

With get128 inlined, the compiler could optimize the code as a whole
and reduce the unexpected memory operations, and as a result, the
query performance is improved.