Revert "Merge pull request #52138 from azat/decompressor-inode"
This reverts commit cad5f8ec80de719a2803cf83112a053a2a442ed7, reversing
changes made to 1d2955d2e490c86953f4506085a6b18375f3929c.

This was not a good idea, since the underlying problem was that
`/proc/self/exe` was pointing to `qemu-$ARCH-static` (because the code
uses realpath() over normal interface readlink(), which is not caught by
the qemu linux-user).

And this means that later, it will try to overwrite incorrect binary and
then execute some garbage.

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>