tests: use explicit address for console to avoid EADDRINUSE
The problem is that MinIO uses dynamic port for console address by
default, which may be below ip_local_port_range, and in this case it ca
be reused in a short time, sicne first MinIO allocates the socket, then
close this socket, and only after try to bind to it.

And even though this is a problem of MinIO I'm not a go developer to fix
it.

v2: use long notation of the 127.0.0.1 (that version of MinIO on CI
cannot handle 127.1, while 2023-07-21T21-12-44Z can)

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>