Optimize DB::combineFilters with AVX512_VBMI2 intrinsic
The DB::combineFilters combines two UInt8 columns (filters) by
scanning the first column for non-zero elements and replacing them
with elements in the second column contiguously.

This function is optimized with Intel's AVX512_VBMI2 intrinsic,
_mm512_mask_expandloadu_epi8, which is designed for implementing
the exact operation when the bits in the writemask are set when
the corresponding packed integers in the first filter are non-zero,
and are not set otherwise. The dynamic dispatch ensures that this
optimization only applies to devices with AVX512_VBMI2 support.