Parallel clone sparse/shallow submodules
(buzzword overload in the title, sorry)

Script update-submodules.sh does a sparse and shallow checkout of
submodules. It is offered as convenience for developers who want to
avoid long checkout times and save disk space. The script is called by a
few build jobs in CI to check that it doesn't break. This PR
parallelizes the actual checkout which reduces script runtime from 425
sec to 258 sec.