Set log_comment to the file name while processing files in client
This will be useful for fuzzer, to know which file had been processed to
trigger the crash, since right now you need to find unique parts of the
query that had not been added by fuzzer to reproduce.

But I guess this will be useful not only for fuzzing, but for general
introspection as well.

Signed-off-by: Azat Khuzhin <a.khuzhin@semrush.com>