Add more doc examples for the EXISTS operator
The existing documentation for EXISTS makes it look like it only works
as part of a WHERE clause, when in fact, it can be used in many places.
Changing the docs to note that a bit differently with another example
should provide more clarity on the use of this term.